requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
distro==1.9.0
Django==6.0.5
django-browser-reload==1.21.0
django-tailwind==4.4.2
gunicorn==26.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.17
langchain-core==1.3.3
langchain-openai==1.2.1
langchain-protocol==0.0.15
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.14
langsmith==0.8.2
openai==2.35.1
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.12
pydantic==2.13.4
pydantic_core==2.46.4
pytailwindcss==0.3.0
pytesseract==0.3.13
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
whitenoise==6.12.0
xxhash==3.7.0
zstandard==0.25.0
